Ejemplo del sistema de velocidades montado.

+++ EjemploSistemaTurnos/Sistema Turnos/Personaje.cs
@@ -8,11 +8,11 @@ namespace Sistema_Turnos
 {
     internal class Personaje
     {
-        public double velocidad; //El atributo de velocidad del personaje. Define dónde empieza en la barra de turnos y donde se coloca al utilizar su acción.
-        public double accion; //El valor de acción en el combate. Define su posición en la barra.
+        public int velocidad; //El atributo de velocidad del personaje. Define dónde empieza en la barra de turnos y donde se coloca al utilizar su acción.
+        public int accion; //El valor de acción en el combate. Define su posición en la barra.
         public string nombre; //El nombre del personaje. Sin más.
 
-        public Personaje(double velocidad, double accion, string nombre) {
+        public Personaje(int velocidad, int accion, string nombre) {
             this.velocidad = velocidad;
             this.nombre = nombre;
             this.accion = accion;
